pipeline/scrapers/requirements.txt
psycopg2-binary>=2.9,<3
requests>=2.31,<3
tldextract>=5.1,<6
pytest>=8.0,<9
